move class to Scores where it's used
